app-python/src/requirements.txt
fastapi
uvicorn[standard]
asgiref
python-dotenv
python-dateutil 
python-multipart 
requests
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
